storages are moved into actors

+++ src/main/scala/Item.scala
@@ -1,5 +1,5 @@
 /**
   * Created by yuriy on 28.10.17.
   */
-final case class Item(id: Long, var name: String)
+final case class Item(id: Int, var name: String)
